move new Superpower-stuff into it's own folder/namespace
